Surfaces can now be grouped into Objects
You can mark a set of Surfaces that belong together as a "thing" or
"light" in the scene description to make them faces of the same
simulated physical entity. This way you can define refractive Things
delimited by Triangles (make sure the mesh is closed though).
If a Surface is defined outside explicit "things" and "lights" it
becomes an Object of its own.
The high-level properties of Surfaces are floated up to Object, Thing
and Light.